pyproject.toml
[project]
name = "rag-pipeline"
version = "2.0.0"
description = "OpenAI API + Qdrant Hybrid Search 기반 RAG Pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    # Vector DB - Qdrant Hybrid Search
    "qdrant-client>=1.12.0",
    "fastembed>=0.4.0",
    # Reranker API Client
    "httpx>=0.27.0",
    # Preprocessing
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "langchain-experimental>=0.3.0",
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # Cache & Database
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ipykernel>=6.29.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
